requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
emoji==2.0.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.28.1
urllib3==1.26.11
yadisk==1.2.15
yarl==1.7.2
